code,middleware: Updated all functionalities+gapbs
This change mostly finalizes most of the necessary changes needed for
the MMIO driver. There is monitor thread for the FAM that constantly
checks for newer updates. The functional version of removal of
permission is also added in this commit but is not tested yet. Other
than that, limitations of the current implementation is also listed in
the code. A bit of cleaning up is required IMO but is not necessary to
start testing functional correctness.

The gapbs middleware is now added to the codebase as well. The purpose
of the gapbs middleware is to manage such shared memory regions at an
application level as the user is not expected to know where the graph is
allocated.

Signed-off-by: Kaustav Goswami <kggoswami@ucdavis.edu>