[1/] wip,middleware,code: middleware for gapbs
This change is the start of a series of patches to make the library
practical for the user to use. This particular middleware is for gapbs.
The version of gapbs this is depended on is darchr:shared-gapbs on
github. It does need a couple of more changes to understand where are
the data located before it becomes fully functional. The idea is that
the user will specify a size to the allocator and the host allocator
will allocate a huge graph. The worker will have limited access to the
graph (defined by the size itself). Entries of each region where the
worker has access will be first acquired when the worker calls for the
mmap.

Signed-off-by: Kaustav Goswami <kggoswami@ucdavis.edu>