code: major overhaul - finalized this version
This change updates the code base to allow cache-line sized metadata for
managing shared disaggregated memory. There are simple APIs that the
user is expected to use to create permissions for the shared memory
space. There is a clean separation of initialization of the memory,
allocation and permission requests. The initial API for the OS now
simply initializes the /dev/dax device. The user then needs to request
permissions to access more parts of the memory by creating entries and
pushing them to the shared memory. The FAM merges most the entries thus
optimizing the overall storage and preventing copies of the permission
table.

Signed-off-by: Kaustav Goswami <kggoswami@ucdavis.edu>